requirements.txt
# LangChain 相关
langchain-openai>=0.1.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-tavily>=0.1.0
langgraph>=0.2.0
langgraph-checkpoint>=1.0.0
langgraph-checkpoint-sqlite>=1.0.0
# 向量数据库
chromadb>=0.4.0
# Web 框架
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.30.0
# 数据处理
pydantic>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
# 中文分词
jieba>=0.42.0
# Embedding 和 Rerank 模型
sentence-transformers>=2.2.0
# HTTP 客户端
requests>=2.31.0
openai>=1.0.0
